fix: ensure our capf comes before others
Our capf is very specific (only activating on the first line, and only when
either there is a single word or we're matching inside brackets for scope. It
can therefore come early in the matching list, allowing it to activate before
other capf's that may match in many circumstances (e.g. cape-dabbrev)

Ref: #5